image upload and download

+++ controllers/admin.js
@@ -1,7 +1,9 @@
 const { validationResult } = require('express-validator/check');
 
 const mongoose = require('mongoose');
+
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   const message = req.flash('error');
@@ -22,9 +24,27 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      errorMessage: 'Attached file is not an image',
+      oldInput: {
+        title: title,
+        price: price,
+        description: description,
+      },
+      validationErrors: [],
+    });
+  }
+
+  const imageUrl = image.path;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -45,7 +65,6 @@ exports.postAddProduct = (req, res, next) => {
   }
 
   const product = new Product({
-    _id: new mongoose.Types.ObjectId('5ec63b581545d93040f8df7e'),
     title: title,
     price: price,
     imageUrl: imageUrl,
@@ -97,8 +116,9 @@ exports.postEditProduct = (req, res, next) => {
   prodId = req.body.productId;
   const upTitle = req.body.title;
   const upPrice = req.body.price;
-  const upImageUrl = req.body.imageUrl;
+  const image = req.file;
   const upDescription = req.body.description;
+
   const errors = validationResult(req);
 
   if (!errors.isEmpty()) {
@@ -110,7 +130,6 @@ exports.postEditProduct = (req, res, next) => {
       product: {
         title: upTitle,
         price: upPrice,
-        imageUrl: upImageUrl,
         description: upDescription,
         _id: prodId,
       },
@@ -128,7 +147,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = upTitle;
       product.price = upPrice;
       product.description = upDescription;
-      product.imageUrl = upImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       return product.save().then((result) => {
         console.log('UPDATED PRODUCT');
         res.redirect('/admin/products');
@@ -143,7 +165,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found.'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
